+++ projects/ng-select2-component/src/lib/select2.module.ts
@@ -1,6 +1,9 @@
+import { OverlayModule } from '@angular/cdk/overlay';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+import { InfiniteScrollModule } from 'ngx-infinite-scroll';
 
 import { Select2Hint } from './select2-hint.component';
 import { Select2Label } from './select2-label.component';
@@ -10,7 +13,9 @@ import { Select2 } from './select2.component';
     imports: [
         CommonModule,
         FormsModule,
-        ReactiveFormsModule
+        OverlayModule,
+        ReactiveFormsModule,
+        InfiniteScrollModule
     ],
     declarations: [
         Select2Hint,